# openapi.yaml
openapi: 3.0.3
info:
  title: JobMajunga (Efiploi) API
  version: 1.0.0
  description: API REST centralisée (Node.js/Express/MySQL) pour WPF + Flutter.
servers:
  - url: https://api.jobmajunga.local/v1
    description: Local/Dev (exemple)
...
security:
  - bearerAuth: []
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      security: []
      summary: Liste offres (publique)
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Recherche texte
        - in: query
          name: status
          schema:
            type: string
            enum: [published]
          description: Public = published uniquement
        - in: query
          name: contractType
          schema:
            type: string
            enum: [CDI, CDD, Freelance, Stage, Alternance]
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedJobOffers"
    post:
      tags: [Jobs]
      summary: Créer une offre (recruteur)
      responses:
        "201":
          description: Offre créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOfferCreate"
  /jobs/mine:
    get:
      tags: [Jobs]
      summary: Liste des offres appartenant au recruteur connecté (ou toutes pour admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, pending_approval, published, expired, archived]
      responses:
        "200":
          description: Offres du recruteur/admin
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/JobOffer" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /jobs/manage/{id}:
    get:
      tags: [Jobs]
      summary: Détail d’une offre pour gestion interne (recruteur propriétaire ou admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Offre (tous statuts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}:
    get:
      tags: [Jobs]
      security: []
      summary: Détail offre (publique si publiée)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Offre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Jobs]
      summary: Modifier une offre (recruteur propriétaire ou admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOfferUpdate"
      responses:
        "200":
          description: Offre mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Jobs]
      summary: Supprimer une offre (recruteur propriétaire ou admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Supprimée }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/duplicate:
    post:
      tags: [Jobs]
      summary: Dupliquer une offre (recruteur)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "201":
          description: Offre dupliquée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/submit:
    post:
      tags: [Jobs]
      summary: Soumettre une offre à approbation (recruteur)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Statut mis à pending_approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/approve:
    post:
      tags: [Jobs]
      summary: Approuver et publier (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Publiée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/archive:
    post:
      tags: [Jobs]
      summary: Archiver une offre (recruteur propriétaire ou admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Archivée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOffer"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/view:
    post:
      tags: [Jobs]
      security: []
      summary: Incrémenter compteur de vues (publique)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: OK }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Forbidden:
      description: Accès interdit (RBAC)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string, nullable: true }
...
    JobOffer:
      type: object
      required: [id, recruiterId, title, description, contractType, status, viewsCount, createdAt, updatedAt]
      properties:
        id: { type: integer }
        recruiterId: { type: integer }
        title: { type: string }
        description: { type: string }
        contractType:
          type: string
          enum: [CDI, CDD, Freelance, Stage, Alternance]
        location: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        salaryMin: { type: number, nullable: true }
        salaryMax: { type: number, nullable: true }
        category: { type: string, nullable: true }
        skills:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [draft, pending_approval, published, expired, archived]
        viewsCount: { type: integer, minimum: 0 }
        expiresAt: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobOfferCreate:
      type: object
      required: [title, description, contractType]
      properties:
        title: { type: string }
        description: { type: string }
        contractType:
          type: string
          enum: [CDI, CDD, Freelance, Stage, Alternance]
        location: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        salaryMin: { type: number, nullable: true }
        salaryMax: { type: number, nullable: true }
        category: { type: string, nullable: true }
        skills:
          type: array
          items: { type: string }
        expiresAt: { type: string, format: date, nullable: true }
    JobOfferUpdate:
      allOf:
        - $ref: "#/components/schemas/JobOfferCreate"
        - type: object
          properties:
            status:
              type: string
              enum: [draft, pending_approval, published, expired, archived]
    PaginatedJobOffers:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/JobOffer" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...